Document faction authoring component states
